Pyplis cell calibration almost ready
Pyplis version of cell calibration is now almost completed. I have updated the previous calibration procedure to save the dark-corrected images so that we can read these in and then the pyplis procedure can be used with darkcorr=False and it should still give valid results. 

TODO test this function and then continue the TODOs from previous commit