Gas calibration finished
Updated ppm.m to cm-2 conversion - it was wrong before.
Updated cropping of images for cell calibration. Should work correctly now.
Off-band image fits are still very wrong, possibly because the initial guess is inaccurate for a non-absorber, might get stuck in a local minimum. But I have inspected this and it shouldn't affect the final result of the AA fit.