Created disk space checking script
New script checks disk space used by image and spectra data and deletes files (oldest first) if data size exceeds a threshold. Currently defining the threshold in the script, but could be defined in a separate text file to provide external access. Default is 10GB.
The script closes once the threshold is reached - it is not an endless loop.

TODO
> Need to add this file to the scheduler - perhaps run it every 5 minutes or so?