LDF complete
LDF correction for spectrometer complete and options for editing are all available in the GUI.

Added time series plot for columns densities from spectrometer, this plot LDF of spectrometer simultaneously. All seems to be working well. NOTE: LDF calculation can be set to occur every x minutes. But if the previous calibration retrieved an LDF of NaN, we will continue to evaluate spectra for LDF until we get a valid answer.