Added a comms receiver handler for GUI
Communications that are received form the instrument are now then used by GUICommRecvHandler which acts on the responses. The idea is that it will print messages to a message window. At the moment it doesn't do much. It have made it update acquisition settings retrieved from {LOG: 1}. Not tested it!