Created message box
Now have a message box window. It may want optimising and it hasn't been tested yet, but it was copied from old pycam code so should work fine. Also it is cut off at the bottom of the screen so may need to put the entire camera window in a scroll frame so can see the bottom.