Load/save settings working
Fully functional load/save settings of camera and spectrometer on start-up/shut-down.

Note SSA, SSB, SSS will not change if Auto shutter speed is enabled. May need to send 2 rounds of commands, on to shut of Auto SS and the second to set shutter speed - if done in one round, if the SSA command is applied first, it will fail as AutoSS still won't have been turned off yet.