Image registration bug fix
Bug where the post-process sequence directory would be reloaded whenever image registration was changed has been fixed. This didn't work when changing image registration whilst acquiring images, as it made caused the system to reload the post-process directory so Villarrica images kept showing up.

Also made it so that on the end of a calibration, the cell calibration frame is updated and displayed immediately.