Big tidy up
No huge changes, just tidying files.

Small changes to controllers, so there is a timeout on setting shutter speed for camera (so it doesn't jam up) which is now used for dark acquisitions too.

Maximum camera shutter speed is now 1 second, not 1.5 as 1.5s leads to too many dropped frames