check_run.py now won't interfere with manual capture
Fix issue #9. There was a bug that meant check_run.py probably shut down the instrument when it wasn't acquiring, even if the user was interacting with the software - e.g. the user wanted to do manual acquisitions or just update the system whilst not in automated capture.

Changes should fix this, although it hasn't been checked on an instrument (don't currently have one)