Plotting calibration parameters fixed
First step in addressing #16.

A simple fix that means that the first of the calibration fit coefficients is plotted in time series and the fit error provided by pyplis is plotted on the same axis. These data also need saving (not currently done), and this won't work correctly for second-order polynomial fits as we will only be plotting the x^2 term.