GUI connection to instrument now cycles through ports
The GUI used to rely on the user changing the port for external connections if the first port fails (the pycam software on the pi can use different port numbers to setup the external connection depending on what ports are available). Now when clicking connect on the GUI it will just cycle through each port if the first one fails.

NOTE: not tested yet..

TODO: If this all works I need to then remove the option to change ports in the GUI