SR4 spectrometer compatibility update
Some integration times aren't compatible with the SR4 so have been removed - they weren't critical times as they were too short or long for skylight measurements anyway.

Also required an update to use seabreeze's pyseabreeze backend. Not totally clear on the ramifications of this, but it seems to work in basic tests.