Making preloaded files more compatible
I've found that editing the preloaded calibration csv file in excel then saving it seems to append commas to the end of all header lines. This interferes with the loading. Just made a small update to make PyplisWorker able to deal with extra commas at the end of the header file by splitting them out before converting to int.