Fixes #111 and a dark spectrum bug
Fixed issue #111 that  meant the default spectra directory was processed even if a new directory had been loaded, if the settings GUI was opened after loading the new spectra directory.

I've also fixed a bug that meant that changing the dark spectra directory didn't correctly update the plot or processing for the first spectrum. NOTE this was a minor bug that wouldn't affect processing once "Process DOAS" was clicked - it was only an issue for the first spectrum when it is processed automatically on directory load.