Fixes #142
Fixes #142 by setting the ILS frame flag to False when the calibration frame is closed down.

This also simultaneously fixed a bug I found that meant you could open an unlimited number of calibration frames simultaneously. Now instead only one can be opened - the frame is brought back to the front if hidden and a new attempt to open it is made.