Checkbutton state configured
Got a rough working version of the checkbutton state being disabled when certain flows aren't present.

This isn't perfect as it doesn't update immediately when you re-enable a flow-type. But clicking the plot again or starting the analysis will make that state active again so things do work in a roundabout way.